Change how we represent criterion values.
KodKod represents an n-ary relation as a matrix of size N^n where N is
the number of atoms in the universe. Since this matrix is a flat array,
its size is limited to 2^32 elements. Thus if we have a 3-ary relation
such as the mapping between contractor methods and value units to metric
values, we cannot have a bit width greater than 10 because the number of
integer atoms will cause N^3 to exceed 2^32.

Thus we cannot have 3-ary relations so we need to express the values as
facts instead of another relation.