Revert "Updated bitwidths for Search and Rescue Problems to be large enough."
This reverts commit 6a45f57d3735f9018147c7f29a1a54b686114013.

This is only needed for the Clafer Moo universe, and it makes the
problems way too large for solvers not using int-elision. This makes the
modified models useless for comparison purposes.

Thus since we don't care about the Clager Moo universe we will go back
to the original bit widths.